# packages/api-client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type CreateUrlDto = {
    /**
     * The URL to shorten
     */
    url: string;
    /**
     * The slug of the URL
     */
    slug?: string;
};

export type Url = {
    /**
     * The ID of the URL
     */
    id: string;
    /**
     * The URL
     */
    url: string;
    /**
     * The slug of the URL
     */
    slug: string;
    /**
     * The date and time the URL was created
     */
    createdAt: string;
    /**
     * The date and time the URL was updated
     */
    updatedAt: string;
    /**
     * The date and time the URL was deleted
     */
    deletedAt?: string;
    /**
     * The ID of the user who created the URL
     */
    userId?: string;
};

export type UpdateSlugDto = {
    /**
     * The slug of the URL
     */
    slug: string;
};

export type User = {
    [key: string]: unknown;
};

export type UrlTracking = {
    /**
     * The ID of the URL tracking
     */
    id: string;
    /**
     * The ID of the URL
     */
    urlId: string;
    /**
     * The referrer domain
     */
    referrerDomain: string;
    /**
     * The browser
     */
    browser: string;
    /**
     * The operating system
     */
    operatingSystem: string;
    /**
     * The device type
     */
    deviceType: string;
    /**
     * The language
     */
    language: string;
    /**
     * The IP address
     */
    ipAddress: string;
    /**
     * The date and time the URL tracking was created
     */
    createdAt: string;
    /**
     * The date and time the URL tracking was created
     */
    updatedAt: string;
};

export type GetHealthData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health';
};

export type GetHealthResponses = {
    200: unknown;
};

export type GetUrlsByUserData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/urls';
};

export type GetUrlsByUserResponses = {
    /**
     * The URLs from the logged in user
     */
    200: Array<Url>;
};

export type GetUrlsByUserResponse = GetUrlsByUserResponses[keyof GetUrlsByUserResponses];

export type CreateUrlData = {
    /**
     * The URL to shorten
     */
    body: CreateUrlDto;
    path?: never;
    query?: never;
    url: '/urls';
};

export type CreateUrlResponses = {
    /**
     * The shortened URL
     */
    200: Url;
};

export type CreateUrlResponse = CreateUrlResponses[keyof CreateUrlResponses];

export type GetUrlBySlugData = {
    body?: never;
    path: {
        slug: string;
    };
    query?: never;
    url: '/urls/{slug}';
};

export type GetUrlBySlugResponses = {
    /**
     * The URL properties
     */
    200: Url;
};

export type GetUrlBySlugResponse = GetUrlBySlugResponses[keyof GetUrlBySlugResponses];

export type DeleteUrlData = {
    body?: never;
    path: {
        id: string;
    };
    query?: never;
    url: '/urls/{id}';
};

export type DeleteUrlResponses = {
    /**
     * The URL was deleted
     */
    200: {
        success?: boolean;
    };
};

export type DeleteUrlResponse = DeleteUrlResponses[keyof DeleteUrlResponses];

export type UpdateSlugData = {
    body: UpdateSlugDto;
    path: {
        id: string;
    };
    query?: never;
    url: '/urls/{id}/slug';
};

export type UpdateSlugResponses = {
    /**
     * The URL properties updated
     */
    200: Url;
};

export type UpdateSlugResponse = UpdateSlugResponses[keyof UpdateSlugResponses];

export type GoogleAuthData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/auth/oauth/google';
};

export type GoogleAuthResponses = {
    200: unknown;
};

export type GoogleAuthCallbackData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/auth/oauth/google/callback';
};

export type GoogleAuthCallbackResponses = {
    200: unknown;
};

export type AuthenticateWithRefreshTokenData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/auth/refresh';
};

export type AuthenticateWithRefreshTokenResponses = {
    200: unknown;
};

export type SignoutData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/auth/signout';
};

export type SignoutResponses = {
    200: unknown;
};

export type GetUserProfileData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/users/profile';
};

export type GetUserProfileResponses = {
    /**
     * The user profile
     */
    200: User;
};

export type GetUserProfileResponse = GetUserProfileResponses[keyof GetUserProfileResponses];

export type GetUrlAnalyticsData = {
    body?: never;
    path: {
        urlId: string;
    };
    query: {
        from: string;
        to: string;
    };
    url: '/urls/trackings/{urlId}';
};

export type GetUrlAnalyticsResponses = {
    /**
     * The URL tracking visits for the given time range
     */
    200: {
        success?: boolean;
        data?: Array<UrlTracking>;
    };
};

export type GetUrlAnalyticsResponse = GetUrlAnalyticsResponses[keyof GetUrlAnalyticsResponses];

export type TrackUrlData = {
    body?: never;
    path: {
        urlId: string;
    };
    query?: never;
    url: '/urls/trackings/{urlId}';
};

export type TrackUrlResponses = {
    /**
     * The URL tracking was created
     */
    200: UrlTracking;
};

export type TrackUrlResponse = TrackUrlResponses[keyof TrackUrlResponses];

export type ClientOptions = {
    baseUrl: string;
};